legacy/requirements.txt
google-genai
Pillow
pillow-heif
python-dotenv
streamlit
